// src/@types/models/product-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { BrandDto } from './brand-dto';
import { CommentDto } from './comment-dto';
import { FavoriteDto } from './favorite-dto';
import { MediaDto } from './media-dto';
import { ProductAttributeDto } from './product-attribute-dto';
import { ProductCategoryDto } from './product-category-dto';
import { RatingDto } from './rating-dto';
import { SupplierProductDto } from './supplier-product-dto';
import { VendorProductDto } from './vendor-product-dto';
/**
 *
 *
 * @export
 * @interface ProductDto
 */
export interface ProductDto {
  /**
   * @type {string}
   * @memberof ProductDto
   */
  id: string;

  /**
   * @type {Date}
   * @memberof ProductDto
   */
  createdAt: Date;

  /**
   * @type {Date}
   * @memberof ProductDto
   */
  updatedAt: Date;

  /**
   * @type {string}
   * @memberof ProductDto
   */
  title: string;

  /**
   * @type {string}
   * @memberof ProductDto
   */
  saleType?: ProductDtoSaleTypeEnum;

  /**
   * @type {string}
   * @memberof ProductDto
   */
  description: string;

  /**
   * @type {BrandDto}
   * @memberof ProductDto
   */
  brand: BrandDto;

  /**
   * @type {string}
   * @memberof ProductDto
   */
  slug: string;

  /**
   * @type {ProductAttributeDto}
   * @memberof ProductDto
   */
  attributes: ProductAttributeDto;

  /**
   * @type {Array<MediaDto>}
   * @memberof ProductDto
   */
  media: Array<MediaDto>;

  /**
   * @type {Array<RatingDto>}
   * @memberof ProductDto
   */
  ratings: Array<RatingDto>;

  /**
   * @type {Array<CommentDto>}
   * @memberof ProductDto
   */
  comments: Array<CommentDto>;

  /**
   * @type {Array<FavoriteDto>}
   * @memberof ProductDto
   */
  favorites: Array<FavoriteDto>;

  /**
   * @type {Array<VendorProductDto>}
   * @memberof ProductDto
   */
  vendorProducts: Array<VendorProductDto>;

  /**
   * @type {Array<SupplierProductDto>}
   * @memberof ProductDto
   */
  supplierProducts: Array<SupplierProductDto>;

  /**
   * @type {Array<ProductCategoryDto>}
   * @memberof ProductDto
   */
  productCategories: Array<ProductCategoryDto>;

  /**
   * @type {number}
   * @memberof ProductDto
   */
  price: number;

  /**
   * @type {number}
   * @memberof ProductDto
   */
  saleCount?: number;

  /**
   * @type {number}
   * @memberof ProductDto
   */
  view?: number;

  /**
   * @type {number}
   * @memberof ProductDto
   */
  rate: number;

  /**
   * @type {boolean}
   * @memberof ProductDto
   */
  status: number;
}

/**
 * @export
 * @enum {string}
 */
export enum ProductDtoSaleTypeEnum {
  PERCENTAGE = 'PERCENTAGE',
  AMOUNT = 'AMOUNT',
}